vue2的capture
